PMDS/requirements.txt
Flask
flask-httpauth
PyQt5
requests
pyyaml
influxdb_client

